requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
pydantic>=2.7.4
jinja2>=3.1.4
python-dotenv>=1.0.1
tqdm>=4.66.4
google-genai>=0.3.0
openai>=1.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
